fix: training label as float32

+++ src/nn/training.py
@@ -30,7 +30,7 @@ def training(n_epochs, train_loader, valid_loader, model, criterion, optimizer, 
         model.eval()
         for data, label in tqdm(valid_loader):
             data = data.to(device=device, dtype=torch.float32)
-            label = label.to(device=device, dtype=torch.long)
+            label = label.to(device=device, dtype=torch.float32)
             with torch.no_grad():
                 output = model(data)
             loss = criterion(output, label)
@@ -52,4 +52,4 @@ def training(n_epochs, train_loader, valid_loader, model, criterion, optimizer, 
             torch.save(model.state_dict(), save_path)
             valid_loss_min = valid_loss
       
-    return train_losses, valid_losses  
+    return train_losses, valid_losses
